requirements.txt
discord.py
functions.py